+++ src/machinable/config/__init__.py
@@ -1,1 +1,0 @@
-from machinable.config.config import Config